Move reconnection code to Connection class
And avoid calling `tearDown` twice
